+++ src/components/card.rs
@@ -1,9 +1,22 @@
+use super::row_scroller::{Scroller, ScrollerProps};
+use super::social_button::{Button, ButtonProps};
 use stylist::yew::styled_component;
 use yew::prelude::*;
 
+#[derive(Properties, PartialEq)]
+pub struct Props {
+    pub title: String,
+    pub button: ButtonProps,
+    pub scroller: ScrollerProps,
+}
+
 #[styled_component]
-pub fn Card() -> Html {
-    let title = "My title name".to_string();
+pub fn Card(props: &Props) -> Html {
+    let Props {
+        title,
+        button,
+        scroller,
+    } = props;
 
     html! {
     <div>
@@ -27,17 +40,7 @@ pub fn Card() -> Html {
                     align-items: center;
                     justify-content: left;
                 "#)}>
-                <button class={css!(r#"
-                        background: none;
-                        border: none;
-                        cursor: pointer;
-                        box-shadow: 0 0 5px 1px rgba(0, 0, 0, 0.7);
-                    "#)}>
-                    <svg width="64" height="64" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-                        <path d="M12 2C6.48 2 2 6.48 2 12C2 17.52 6.48 22 12 22C17.52 22 22 17.52 22 12C22 6.48 17.52 2 12 2ZM12 20C7.59 20 4 16.41 4 12C4 7.59 7.59 4 12 4C16.41 4 20 7.59 20 12C20 16.41 16.41 20 12 20Z" fill="currentColor"/>
-                        <path d="M11 11H13V16H11V11ZM11 7H13V9H11V7Z" fill="currentColor"/>
-                    </svg>
-                </button>
+                <Button ..button.clone() />
                 <h2 class={css!(r#"
                     margin-left: 20px;
                 "#)}>{ title }</h2>
@@ -50,72 +53,9 @@ pub fn Card() -> Html {
                     width: 100%;
                     overflow: hidden;
                 "#)}>
-                <Scroller />
+                <Scroller ..scroller.clone() />
             </div>
         </div>
     </div>
     }
-}
-
-#[styled_component]
-pub fn Scroller() -> Html {
-    let rows: Vec<String> = (1..15).map(|i| format!("Item {i}")).collect();
-
-    html! {
-        <div class={css!(r#"
-            width: 100%;
-            height: 100%;
-            border-radius: 10px;
-
-            background: black;
-
-            padding: 15px;
-            box-sizing: border-box;
-
-            box-shadow: 0 0 5px 1px rgba(0, 0, 0, 0.7);
-            color: white;
-
-            display: flex;
-            flex-direction: column;
-
-            overflow-y: auto;
-            overflow-x: hidden;
-            box-sizing: border-box;
-
-            /* Hide scrollbar for Webkit browsers (Chrome, Safari) */
-            &::-webkit-scrollbar {
-                display: none;
-            }
-
-            /* Hide scrollbar for Firefox */
-            scrollbar-width: none;
-            -ms-overflow-style: none; /* Internet Explorer 10+ */
-        "#)}>
-            { for rows.iter().map(|row| html! { <Row value={row.clone()} /> }) }
-        </div>
-    }
-}
-
-#[styled_component]
-pub fn Row(props: &Props) -> Html {
-    let Props { value } = props;
-
-    html! {
-        <div class={css!(r#"
-            width: 100%;
-            display: flex;
-            justify-content: space-between;
-            padding: 10px;
-            border-bottom: 1px solid white;
-            margin-bottom: 5px;
-            box-sizing: border-box;
-        "#)}>
-            <span>{ value }</span>
-        </div>
-    }
-}
-
-#[derive(Properties, PartialEq)]
-pub struct Props {
-    pub value: String,
 }
